add terraform-side validation of some scrape job properties
