+++ docs/data-sources/cloud_provider_aws_cloudwatch_scrape_jobs.md
@@ -31,7 +31,7 @@ data "grafana_cloud_provider_aws_cloudwatch_scrape_jobs" "test" {
 
 ### Optional
 
-- `scrape_job` (Block Set) One or more AWS CloudWatch Scrape Job blocks associated with the given StackID. (see [below for nested schema](#nestedblock--scrape_job))
+- `scrape_job` (Block List) A list of AWS CloudWatch Scrape Job objects associated with the given StackID. (see [below for nested schema](#nestedblock--scrape_job))
 
 ### Read-Only
 
@@ -51,7 +51,7 @@ Read-Only:
 - `enabled` (Boolean) Whether the CloudWatch Scrape Job is enabled or not.
 - `id` (String) The Terraform Resource ID. This has the format "{{ stack_id }}:{{ job_name }}".
 - `regions` (Set of String) A set of AWS region names that this CloudWatch Scrape Job applies to.
-- `service_configuration` (Block Set) Each block dictates what this CloudWatch Scrape Job should scrape for the specified AWS service. (see [below for nested schema](#nestedblock--scrape_job--service_configuration))
+- `service_configuration` (Block List) One or more configuration blocks to dictate what this CloudWatch Scrape Job should scrape. Each block must have a distinct `name` attribute. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--scrape_job--service_configuration))
 
 <a id="nestedblock--scrape_job--service_configuration"></a>
 ### Nested Schema for `scrape_job.service_configuration`
@@ -59,9 +59,9 @@ Read-Only:
 Read-Only:
 
 - `is_custom_namespace` (Boolean) Whether the service name is a custom, user-generated metrics namespace, as opposed to a standard AWS service metrics namespace.
-- `metric` (Block Set) Each block configures a metric and their statistics to scrape. (see [below for nested schema](#nestedblock--scrape_job--service_configuration--metric))
+- `metric` (Block List) One or more configuration blocks to configure metrics and their statistics to scrape. Each block must represent a distinct metric name. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--scrape_job--service_configuration--metric))
 - `name` (String) The name of the service to scrape. See https://grafana.com/docs/grafana-cloud/monitor-infrastructure/aws/cloudwatch-metrics/services/ for supported services, metrics, and their statistics.
-- `resource_discovery_tag_filter` (Block Set) Each block configures a tag filter applied to discovery of resource entities in the associated AWS account. (see [below for nested schema](#nestedblock--scrape_job--service_configuration--resource_discovery_tag_filter))
+- `resource_discovery_tag_filter` (Block List) One or more configuration blocks to configure tag filters applied to discovery of resource entities in the associated AWS account. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--scrape_job--service_configuration--resource_discovery_tag_filter))
 - `scrape_interval_seconds` (Number) The interval in seconds to scrape the service. See https://grafana.com/docs/grafana-cloud/monitor-infrastructure/aws/cloudwatch-metrics/services/ for supported scrape intervals.
 - `tags_to_add_to_metrics` (Set of String) A set of tags to add to all metrics exported by this scrape job, for use in PromQL queries.
