+++ docs/resources/cloud_provider_aws_cloudwatch_scrape_job.md
@@ -122,7 +122,7 @@ resource "grafana_cloud_provider_aws_cloudwatch_scrape_job" "test" {
 ### Optional
 
 - `enabled` (Boolean) Whether the CloudWatch Scrape Job is enabled or not.
-- `service_configuration` (Block Set) Each block dictates what this CloudWatch Scrape Job should scrape for the specified AWS service. (see [below for nested schema](#nestedblock--service_configuration))
+- `service_configuration` (Block List) One or more configuration blocks to dictate what this CloudWatch Scrape Job should scrape. Each block must have a distinct `name` attribute. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--service_configuration))
 
 ### Read-Only
 
@@ -138,8 +138,8 @@ Required:
 Optional:
 
 - `is_custom_namespace` (Boolean) Whether the service name is a custom, user-generated metrics namespace, as opposed to a standard AWS service metrics namespace.
-- `metric` (Block Set) Each block configures a metric and their statistics to scrape. (see [below for nested schema](#nestedblock--service_configuration--metric))
-- `resource_discovery_tag_filter` (Block Set) Each block configures a tag filter applied to discovery of resource entities in the associated AWS account. (see [below for nested schema](#nestedblock--service_configuration--resource_discovery_tag_filter))
+- `metric` (Block List) One or more configuration blocks to configure metrics and their statistics to scrape. Each block must represent a distinct metric name. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--service_configuration--metric))
+- `resource_discovery_tag_filter` (Block List) One or more configuration blocks to configure tag filters applied to discovery of resource entities in the associated AWS account. When accessing this as an attribute reference, it is a list of objects. (see [below for nested schema](#nestedblock--service_configuration--resource_discovery_tag_filter))
 - `scrape_interval_seconds` (Number) The interval in seconds to scrape the service. See https://grafana.com/docs/grafana-cloud/monitor-infrastructure/aws/cloudwatch-metrics/services/ for supported scrape intervals.
 - `tags_to_add_to_metrics` (Set of String) A set of tags to add to all metrics exported by this scrape job, for use in PromQL queries.
