+++ internal/resources/cloudprovider/data_source_aws_cloudwatch_scrape_job.go
@@ -40,8 +40,8 @@ var (
 			},
 		},
 		Blocks: map[string]schema.Block{
-			"service_configuration": schema.SetNestedBlock{
-				Description: "Each block dictates what this CloudWatch Scrape Job should scrape for the specified AWS service.",
+			"service_configuration": schema.ListNestedBlock{
+				Description: "One or more configuration blocks to dictate what this CloudWatch Scrape Job should scrape. Each block must have a distinct `name` attribute. When accessing this as an attribute reference, it is a list of objects.",
 				NestedObject: schema.NestedBlockObject{
 					Attributes: map[string]schema.Attribute{
 						"name": schema.StringAttribute{
@@ -63,8 +63,8 @@ var (
 						},
 					},
 					Blocks: map[string]schema.Block{
-						"metric": schema.SetNestedBlock{
-							Description: "Each block configures a metric and their statistics to scrape.",
+						"metric": schema.ListNestedBlock{
+							Description: "One or more configuration blocks to configure metrics and their statistics to scrape. Each block must represent a distinct metric name. When accessing this as an attribute reference, it is a list of objects.",
 							NestedObject: schema.NestedBlockObject{
 								Attributes: map[string]schema.Attribute{
 									"name": schema.StringAttribute{
@@ -79,8 +79,8 @@ var (
 								},
 							},
 						},
-						"resource_discovery_tag_filter": schema.SetNestedBlock{
-							Description: "Each block configures a tag filter applied to discovery of resource entities in the associated AWS account.",
+						"resource_discovery_tag_filter": schema.ListNestedBlock{
+							Description: "One or more configuration blocks to configure tag filters applied to discovery of resource entities in the associated AWS account. When accessing this as an attribute reference, it is a list of objects.",
 							NestedObject: schema.NestedBlockObject{
 								Attributes: map[string]schema.Attribute{
 									"key": schema.StringAttribute{
